+++ admin/map_metric_api.php
@@ -1,0 +1,6 @@
+<?php
+require_once __DIR__ . '/../endpoints/_includes/functions.php';
+header("content-type: application/json");
+echo json_encode(getMapMetrics(
+    $_REQUEST['service_body_id']
+));